fix: 검색 페이지 검색창 이동 및 필터·뷰 토글 UX 수정
